+++ src/actions/types/todolistTypes.ts
@@ -1,2 +1,3 @@
 export const SET_NEW_LIST = 'SET_NEW_LIST';
 export const REMOVE_LIST = 'REMOVE_LIST';
+export const EDIT_LIST = 'EDIT_LIST';